LayoutTests/platform/mac/accessibility/search-predicate-element-count-expected.txt

This tests that search predicate based element count results are accurate.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS elementCount is 2
PASS elementCount is 3
PASS elementCount is 4
PASS elementCount is 5
PASS elementCount is 6
PASS elementCount is 6
PASS elementCount is 4
PASS successfullyParsed is true

TEST COMPLETE
